*** empty log message ***

+++ common/bcdfloat.cc
@@ -215,16 +215,16 @@ BCDFloat::BCDFloat(int4 v)
         d_[0] = v;
         d_[P] = 1;
     }
-    else if (v < BASE*BASE) {
+    else if (v < ((int4)BASE)*BASE) {
         d_[0] = v/BASE;
-        d_[1] = v - d_[0]*BASE;
+        d_[1] = v - d_[0]*((int4)BASE);
         d_[P] = 2;
     }
     else {
-        d_[0] = v/(BASE*BASE);
-        v -= d_[0]*(BASE*BASE);
+        d_[0] = v/(((int4)BASE)*BASE);
+        v -= d_[0]*(((int4)BASE)*BASE);
         d_[1] = v/BASE;
-        d_[2] = v - d_[1]*BASE;
+        d_[2] = v - d_[1]*((int4)BASE);
         d_[P] = 3;
     }
 
@@ -765,7 +765,7 @@ void BCDFloat::mul(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
             ca = 0;
             if (v >= BASE) {
                 ca = v / BASE;
-                v = v - ca*BASE;
+                v = v - ca*((int4)BASE);
             }
             acc.d_[j] = v;
         }
@@ -866,7 +866,7 @@ void BCDFloat::div(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
         BCDFloat acc;
         BCDFloat b1;
 
-        u = BASE/(b->d_[0]+1);
+        u = ((int4)BASE)/(b->d_[0]+1);
 
         if (u != 1) {
             /* prenormialise `a' and move into acc using spare digit */
@@ -876,7 +876,7 @@ void BCDFloat::div(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
                 ca = 0;
                 if (v >= BASE) {
                     ca = v/BASE;
-                    v -= ca*BASE;
+                    v -= ca*((int4)BASE);
                 }
                 acc.d_[i] = v;
             }
@@ -889,7 +889,7 @@ void BCDFloat::div(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
                 ca = 0;
                 if (v >= BASE) {
                     ca = v/BASE;
-                    v -= ca*BASE;
+                    v -= ca*((int4)BASE);
                 }
                 b1.d_[i] = v;
             }
@@ -906,7 +906,7 @@ void BCDFloat::div(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
         for (;;) {
             if (acc.d_[0] == b1.d_[0]) q = BASE-1;
             else {
-                v = acc.d_[0]*BASE + acc.d_[1];
+                v = acc.d_[0]*((int4)BASE) + acc.d_[1];
                 q = v/b1.d_[0];
 
                 while (b1.d_[1]*q > ((v - q*b1.d_[0])*BASE + acc.d_[2])) {
@@ -932,7 +932,7 @@ void BCDFloat::div(const BCDFloat* a, const BCDFloat* b, BCDFloat* c)
                     ca = 0;
                     if (v < 0) {
                         ca = (-v + BASE-1)/BASE;
-                        v += ca*BASE;
+                        v += ca*((int4)BASE);
                     }
                     acc.d_[i] = v;
                 }
@@ -997,7 +997,7 @@ void BCDFloat::mul2(unsigned short* ad, int ea,
             ca = 0;
             if (v >= BASE) {
                 ca = v / BASE;
-                v = v - ca*BASE;
+                v = v - ca*((int4)BASE);
             }
             acc[j] = v;
         }
@@ -1091,7 +1091,7 @@ bool BCDFloat::sqrt(const BCDFloat* a, BCDFloat* r)
                 ca = 0;
                 if (v >= BASE) {
                     ca = v/BASE;
-                    v -= ca*BASE;
+                    v -= ca*((int4)BASE);
                 }
                 t.d_[i] = v;
             }
@@ -1113,10 +1113,10 @@ bool BCDFloat::sqrt(const BCDFloat* a, BCDFloat* r)
 
             q = 0;
             if (ts == as) {
-                q = (((int4) acc.d_[0])*BASE + acc.d_[1])/(((int4) t.d_[0])*BASE+t.d_[1]);
+                q = (acc.d_[0]*((int4)BASE) + acc.d_[1])/(t.d_[0]*((int4)BASE)+t.d_[1]);
             }
             else if (as > ts) {
-                q = (((int4) acc.d_[0])*BASE + acc.d_[1])/t.d_[0];
+                q = (acc.d_[0]*((int4)BASE) + acc.d_[1])/t.d_[0];
             }
 
             if (q) {
@@ -1134,7 +1134,7 @@ bool BCDFloat::sqrt(const BCDFloat* a, BCDFloat* r)
                         ca = 0;
                         if (v >= BASE) {
                             ca = v/BASE;
-                            v -= ca*BASE;
+                            v -= ca*((int4)BASE);
                         }
                         u.d_[i] = v;
                     }
